refactor(schemas): migrate Phase 32 endpoint schemas — ide_integration.py 15 classes + 4 enums (#6042)
- ide_integration.py: migrate DiagnosticSeverity, CodeActionKind, PatternCategory→
  IDEPatternCategory, CompletionItemKind enums + Position→LSPPosition, Range→LSPRange,
  Diagnostic, TextEdit, CodeAction, AnalysisRequest→IDEAnalysisRequest,
  AnalysisResponse→IDEAnalysisResponse, QuickFixRequest, QuickFixResponse,
  HoverRequest, HoverResponse, ConfigurationUpdate→IDEConfigurationUpdate,
  CompletionItem, CompletionRequest, CompletionResponse → schemas_code.py

Renames to avoid collisions with built-in names and existing schemas:
- Position → LSPPosition (too generic)
- Range → LSPRange (collides with builtin range concept)
- PatternCategory → IDEPatternCategory (collides with analytics_pattern_learning)
- AnalysisRequest → IDEAnalysisRequest (anti_pattern, code_intel use AnalysisRequest)
- AnalysisResponse → IDEAnalysisResponse (anti_pattern has AnalysisResponse)
- ConfigurationUpdate → IDEConfigurationUpdate

schemas_code.py: gains Enum import.
